feat: Add fuel type and consumption to vehicles
Introduce new properties to the Vehicle type to store fuel type and consumption details. This will enable more accurate cost calculations for rides and vehicle maintenance tracking.

Also includes updates to the `RideRequest` type to store calculated fuel cost and the addition of a `FuelPrices` type for managing fuel costs in different types.